requirements.txt
-i https://pypi.org/simple
absl-py==1.4.0; python_version >= '3.6'
accelerate==0.22.0; python_full_version >= '3.8.0'
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
astunparse==1.6.3
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
cached-property==1.5.2
cachetools==5.3.1; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
chex==0.1.7; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
cloudpickle==2.2.1; python_version >= '3.6'
contourpy==1.1.0; python_version >= '3.8'
cycler==0.11.0; python_version >= '3.6'
cython==3.0.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
datasets==2.14.4; python_full_version >= '3.8.0'
decorator==5.1.1; python_version >= '3.5'
dill==0.3.7; python_version >= '3.7'
dm-tree==0.1.8
dopamine-rl==4.0.6; python_version >= '3.5' and python_version < '4'
essential-generators==1.0
etils==1.3.0; python_version >= '3.8'
farama-notifications==0.0.4
-e .
filelock==3.12.2; python_version >= '3.7'
flatbuffers==23.5.26
flax==0.7.2
fonttools==4.42.1; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
fsspec[http]==2023.6.0; python_version >= '3.8'
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gin-config==0.5.0
glfw==2.6.2
google-auth==2.22.0; python_version >= '3.6'
google-auth-oauthlib==1.0.0; python_version >= '3.6'
google-pasta==0.2.0
grpcio==1.57.0; python_version >= '3.7'
gym==0.25.2; python_version >= '3.6'
gym-notices==0.0.8
gymnasium==0.29.1; python_version >= '3.8'
gymnasium-robotics==1.2.2; python_version >= '3.7'
h5py==3.9.0; python_version >= '3.8'
huggingface==0.0.1
huggingface-hub==0.16.4; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
imageio==2.31.2; python_version >= '3.7'
importlib-metadata==6.8.0; python_version < '3.10'
importlib-resources==6.0.1; python_version < '3.10'
jax==0.4.13; python_version >= '3.8'
jaxlib==0.4.13; python_version >= '3.8'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
jsonc-parser==1.1.5; python_version >= '3.5'
keras==2.13.1; python_version >= '3.8'
kiwisolver==1.4.5; python_version >= '3.7'
lazy-loader==0.3; python_version >= '3.7'
libclang==16.0.6
lorem-text==2.1; python_version >= '3.5'
markdown==3.4.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.7.2; python_version >= '3.8'
matplotlib-scalebar==0.8.1; python_version ~= '3.7'
mdurl==0.1.2; python_version >= '3.7'
minigrid==2.3.0; python_version >= '3.7'
ml-dtypes==0.2.0; python_version >= '3.7'
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.3; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
multiprocess==0.70.15; python_version >= '3.7'
nest-asyncio==1.5.7; python_version >= '3.5'
networkx==3.1; python_version >= '3.8'
nltk==3.8.1; python_version >= '3.7'
numpy==1.23.5; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
opencv-python==4.8.0.76; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
optax==0.1.7; python_version >= '3.8'
orbax-checkpoint==0.2.3; python_version >= '3.8'
packaging==23.1; python_version >= '3.7'
palettable==3.3.3; python_version >= '3.7'
pandas==2.0.3; python_version >= '3.8'
pettingzoo==1.24.0; python_version >= '3.8'
pillow==10.0.0; python_version >= '3.8'
platformdirs==3.10.0; python_version >= '3.7'
plotly==5.16.1; python_version >= '3.6'
pooch==1.7.0; python_version >= '3.7'
protobuf==4.24.2; python_version >= '3.7'
psutil==5.9.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyarrow==13.0.0; python_version >= '3.8'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pygame==2.5.1; python_version >= '3.6'
pygments==2.16.1; python_version >= '3.7'
pyopengl==3.1.7
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
pywavelets==1.4.1; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.5.2; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
safetensors==0.3.3
scikit-image==0.21.0; python_version >= '3.8'
scikit-learn==1.3.0; python_version >= '3.8'
scipy==1.10.1; python_version < '3.12' and python_version >= '3.8'
seaborn==0.12.2; python_version >= '3.7'
seaborn-image==0.7.0; python_version < '3.11' and python_version >= '3.8'
sentence-transformers==2.2.2; python_full_version >= '3.6.0'
sentencepiece==0.1.99
setuptools==68.1.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.12; python_version >= '3.8'
tenacity==8.2.3; python_version >= '3.7'
tensorboard==2.13.0; python_version >= '3.8'
tensorboard-data-server==0.7.1; python_version >= '3.7'
tensorboardx==2.6.2.2
tensorflow==2.13.0; python_version >= '3.8'
tensorflow-estimator==2.13.0; python_version >= '3.7'
tensorflow-macos==2.13.0; platform_system == 'Darwin' and platform_machine == 'arm64'
tensorflow-probability==0.21.0; python_version >= '3.8'
tensorstore==0.1.41; python_version >= '3.8'
termcolor==2.3.0; python_version >= '3.7'
tf-slim==1.1.0
threadpoolctl==3.2.0; python_version >= '3.8'
tifffile==2023.7.10; python_version >= '3.8'
tokenizers==0.13.3
toolz==0.12.0; python_version >= '3.5'
torch==2.0.1; python_full_version >= '3.8.0'
torch-ac==1.4.0
torchaudio==2.0.2
torchvision==0.15.2; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
transformers==4.32.1; python_full_version >= '3.8.0'
typing-extensions==4.5.0; python_version >= '3.7'
tzdata==2023.3; python_version >= '2'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.7; python_version >= '3.8'
wheel==0.41.2; python_version >= '3.7'
wrapt==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
xxhash==3.3.0; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
zipp==3.16.2; python_version >= '3.8'
